Create README.md
This will be an epic game, first of it's kind as to an actual, playable Superman game. Guaranteed to make us rich. Help me? Teach me so I can learn to program and code so for 1. I can help in this games development and creation,but also learn to program and write my own. And trust me I can come up with 10 more easily. Thank you and God Bless.....
 Stand for something, or fall for everything....👍👍👍👍👍🙏🙏🙏